+++ frontend/src/app/main/data/workspace.cljs
@@ -717,17 +717,18 @@
                                groups-to-delete groups-to-unmask shapes-to-detach
                                shapes-to-reroot shapes-to-deroot shapes-to-unconstraint]
   (let [ordered-indexes (cph/order-by-indexed-shapes objects ids)
-        shapes (map (d/getf objects) ordered-indexes)]
+        shapes (map (d/getf objects) ordered-indexes)
+        parent (get objects parent-id)]
 
     (-> (pcb/empty-changes it page-id)
         (pcb/with-objects objects)
 
         ;; Remove layout-item properties when moving a shape outside a layout
-        (cond-> (not (ctl/any-layout? objects parent-id))
+        (cond-> (not (ctl/any-layout? parent))
           (pcb/update-shapes ordered-indexes ctl/remove-layout-item-data))
 
         ;; Remove the hide in viewer flag
-        (cond-> (and (not= uuid/zero parent-id) (cph/frame-shape? objects parent-id))
+        (cond-> (and (not= uuid/zero parent-id) (cph/frame-shape? parent))
           (pcb/update-shapes ordered-indexes #(cond-> % (cph/frame-shape? %) (assoc :hide-in-viewer true))))
 
         ;; Move the shapes
@@ -759,8 +760,7 @@
         ;; Reset constraints depending on the new parent
         (pcb/update-shapes shapes-to-unconstraint
                            (fn [shape]
-                             (let [parent      (get objects parent-id)
-                                   frame-id    (if (= (:type parent) :frame)
+                             (let [frame-id    (if (= (:type parent) :frame)
                                                  (:id parent)
                                                  (:frame-id parent))
                                    moved-shape (assoc shape
@@ -782,6 +782,10 @@
                                  (ctl/change-v-sizing? (:id parent) objects (:shapes parent))
                                  (assoc :layout-item-v-sizing :fix))
                                parent)))
+
+        ;; If parent locked, lock the added shapes
+        (cond-> (:blocked parent)
+          (pcb/update-shapes ordered-indexes #(assoc % :blocked true)))
 
         ;; Resize parent containers that need to
         (pcb/resize-parents parents))))